bugfix: Start down making the network more configurable
In preparation for making bridged networking possible. Lots to do yet
though in regards to making things that matter (e.g. replicas) pass
into helmfile values somehow so that we don't have 3 replicas for say
rancher or neuvector if we only have 1 worker node. That is a silly
setup to allow unless someone wants to do crazy stuff.