Quick hacks to get a second vip and a few helmfile refactors
Ripped out the hack helm chart and just using the incubator/raw chart
to act as a locum helm chart to install kube-vip against.

Had to also work around helmfiles not letting me patch the helm output
easily, or I'm not getting how to do it based off the examples.

So I yeeted some logic to use the dynamic golang client to just loop
in the background until the service comes up to update the Service to
add the kube-vip cloud ip config needed.

Hopefully I can remove it but if not whatever.